Added dark mode and style elements
