+++ client/src/app/layout/App.tsx
@@ -1,38 +1,32 @@
-import { Container, CssBaseline} from "@mui/material";
-import{useEffect, useState} from "react";
+import { Container, createTheme, CssBaseline, ThemeProvider} from "@mui/material";
+import { useState } from "react";
 import Catalog from "../../features/catalog/Catalog";
-import { Product } from "../models/product";
 import Header from "./Header";
 
 function App() {
-  const[products, setProducts] = useState<Product[]>([]);
-
-  useEffect(() => {
-    fetch('http://localhost:5000/api/products')
-    .then(response => response.json())
-    .then(data => setProducts(data))
-  }, []);
-
-  function addProduct() {
-    setProducts(prevState =>[...prevState, 
-      {
-        id: prevState.length + 101,
-        name:'product' + (prevState.length + 1), 
-        price: (prevState.length * 100) + 199.99,
-        brand: 'some brand',
-        description:'some description',
-        pictureUrl: 'http://picsum.photo/200',
-      }])
+  const[darkMode, setDarkMode] = useState(false);
+  const paletteType = darkMode ? 'dark' : 'light';
+  const theme = createTheme({
+    palette:{
+      mode: paletteType,
+      background: {
+        default: paletteType === 'light' ? '#eaeaea' : '#121212'
+      }
+    }
+  });
+  
+  function handleThemeChange() {
+    setDarkMode(!darkMode);
   }
 
   return (
-    <>
+    <ThemeProvider theme={theme}>
       <CssBaseline/>
-      <Header/>
+      <Header darkMode={darkMode} handleThemeChange={handleThemeChange}/>
       <Container>
-        <Catalog products = {products} addProduct = {addProduct}/>
+        <Catalog/>
       </Container>
-    </>
+    </ThemeProvider>
   );
 }
